refactored after pull request
