+++ src/pages/private/index.js
@@ -1,26 +1,34 @@
 // outsource dependencies
 import React from 'react';
+import { Spinner } from 'reactstrap';
 import { Switch } from 'react-router-dom';
 import { useSelector } from 'react-redux';
 
 // local dependencies
 import User from './user';
 import Homepage from './homepage';
-import { Spinner } from 'reactstrap';
-import { selector } from '../../redux/reducer';
 import PrivateRoute from '../../utils/private-route';
+import { selector as pagesSelector } from '../reducer';
+import { selector as privateSelector } from './reducer';
 import { PRIVATE_HOMEPAGE, PRIVATE_USER } from '../../constants/routes';
 
 function Private () {
-  const root = useSelector(selector);
+  const { user } = useSelector(pagesSelector);
+  const { initialized } = useSelector(privateSelector);
 
   return <>
     {
-      root.pages.private.initialized ? <Switch>
-        <PrivateRoute path={PRIVATE_USER} isUserAuthed={!!root.pagesData.user}
-          user={root.pagesData.user} component={User}/>
-        <PrivateRoute path={PRIVATE_HOMEPAGE} isUserAuthed={!!root.pagesData.user}
-          user={root.pagesData.user} component={Homepage}/>
+      initialized ? <Switch>
+        <PrivateRoute
+          path={PRIVATE_USER}
+          isUserAuthed={!!user}
+          user={user}
+          component={User}/>
+        <PrivateRoute
+          path={PRIVATE_HOMEPAGE}
+          isUserAuthed={!!user}
+          user={user}
+          component={Homepage}/>
       </Switch> : <div><Spinner color="primary" /> private</div>
     }
   </>;
