+++ src/pages/private/user/reducer.js
@@ -1,5 +1,3 @@
-export const SAGA_LOGOUT = 'SAGA_LOGOUT';
-
 export default function userReducer (state={}) {
   return state;
 }
